renamed SHOW_ADDRESSES to VIEW_ADDRESSES To be more consistent with the other permission names
added a getDescriptions method to Permission. then i realized that i can't pass
data into this method because JSP sucks balls. i don't feel like re-typing all
this data later, so i'm leaving it here for now. jeff, got an idea on
how we can get this information out of the system more easily? maybe if
permission was a class (and represented in the db) and not an enum as i originally suggested, this would
be more easily possible because we could store the description in the db. when we want to
i18n things, we could change the actual description in the db to a key
which we do a lookup on during render.

git-svn-id: https://subetha.googlecode.com/svn/trunk@185 68df8246-1725-11de-bcda-2bb9902f8029