Mail.app->subethasmtpd now works
Figured out that the ehlo/helo commands were failing because Mail.app
is sending a crap IPv6 address that can't be resolved, so let's not 
bother doing that lookup as it really isn't necessary for this code to 
work anyway (at least that's what i believe)

01:20:27,605 INFO  [HelloCommand] HELO [jb?7??IPv6:::1] => 250 131.161.241.39

Also, fixed another major bug that was throwing an NPE as a result of Session.reset()
because an array that is initialized during the class init is being nulled
in init(). don't null. just clear().
01:18:27,239 ERROR [STDERR] Exception in thread "Thread-17" 
01:18:27,240 ERROR [STDERR] java.lang.NullPointerException
01:18:27,241 ERROR [STDERR]     at org.subethamail.smtp.session.Session.addRecipient(Session.java:108)
01:18:27,241 ERROR [STDERR]     at org.subethamail.smtp.command.ReceiptCommand.execute(ReceiptCommand.java:44)
01:18:27,241 ERROR [STDERR]     at org.subethamail.smtp.command.CommandLogger.execute(CommandLogger.java:36)
01:18:27,241 ERROR [STDERR]     at org.subethamail.smtp.command.CommandDispatcher.executeCommand(CommandDispatcher.java:45)
01:18:27,241 ERROR [STDERR]     at org.subethamail.smtp.server.SocketHandler.run(SocketHandler.java:83)
01:18:27,242 ERROR [STDERR]     at java.lang.Thread.run(Thread.java:613)


git-svn-id: https://subetha.googlecode.com/svn/trunk@276 68df8246-1725-11de-bcda-2bb9902f8029