fixed a mess of little bugs regarding output of commands
fixed a bug where the bindAddress was not being set properly.

implemented a watchdog thread that will timeout connections
if they haven't been active for 1 minute.

implemented a max connections hard limit (1000).

we still need to have a way to deal with timing out once DATA
starts flowing in. the only way i can see how to do this would 
be to start another thread to deal with delivery. i'm not sure that
this is necessary at this point in time though. maybe if it was an
internet facing server.

git-svn-id: https://subetha.googlecode.com/svn/trunk@353 68df8246-1725-11de-bcda-2bb9902f8029