The problem:
If subetha is behind a proxy such that requests to http://subetha.host.com:80/ get proxied to...

RewriteRule ^/se(.*) http://subetha.host.com:8080/se$1 [P]

When someone goes to:

http://subetha.host.com/se/

A redirect would be issued by the subetha server to:

http://subetha.host.com:8080/se/home.jsp

Thus exposing the 'hidden' proxy port.

This change takes advantage of the Site URL setting to redirect to the actual site url that is configured.

If nothing is configured (ie: needsconfiguration), then the default is to continue with the existing behavior.

One problem with this configuration is that is if there are virtual hosts on subetha server. This would
expose the servers site url. I'm not sure if this is ideal or not, but I'm not sure the best 
way to handle it yet. One idea would be to do a lookup to see if the incoming domain is configured as
a list and then redirect to that list homepage (ie: the list url setting) instead of the site url setting.

git-svn-id: https://subetha.googlecode.com/svn/trunk@920 68df8246-1725-11de-bcda-2bb9902f8029