i had a message that had:
From MAILER-DAEMON

this was causing a NPE cause the From line was found, but since the address didn't have a 
@ in it the MailUtils.getMboxFrom() was returning null.

the fix is to just skip the message since we probably don't want MAILER-DAEMON messages anyway.

the only problem with this approach is that a mbox full of 'incomplete' emails will 
have those messages ignored.

git-svn-id: https://subetha.googlecode.com/svn/trunk@949 68df8246-1725-11de-bcda-2bb9902f8029